+++ packages/solid-router/src/utils.ts
@@ -1,27 +1,5 @@
 import * as Solid from 'solid-js'
 
-export const useLayoutEffect =
-  typeof window !== 'undefined' ? Solid.createEffect : Solid.createEffect
-
-export const usePrevious = (fn: () => boolean) => {
-  return Solid.createMemo(
-    (
-      prev: { current: boolean | null; previous: boolean | null } = {
-        current: null,
-        previous: null,
-      },
-    ) => {
-      const current = fn()
-
-      if (prev.current !== current) {
-        return { previous: prev.current, current }
-      }
-
-      return prev
-    },
-  )
-}
-
 /**
  * React hook to wrap `IntersectionObserver`.
  *