+++ packages/solid-router-devtools/CHANGELOG.md
@@ -47,6 +47,15 @@
 - Updated dependencies [[`a0191af`](https://github.com/TanStack/router/commit/a0191afd21afe0e7571af8b0faab171f62e71db7)]:
   - @tanstack/solid-router@2.0.0-alpha.0
 
+## 1.166.10
+
+### Patch Changes
+
+- Updated dependencies [[`0545239`](https://github.com/TanStack/router/commit/054523900b2ee19308e5a88417dadfc6923afe30)]:
+  - @tanstack/router-devtools-core@1.167.0
+  - @tanstack/solid-router@1.168.0
+  - @tanstack/router-core@1.168.0
+
 ## 1.166.9
 
 ### Patch Changes