+++ packages/solid-start-client/CHANGELOG.md
@@ -61,6 +61,24 @@
 - Updated dependencies [[`a0191af`](https://github.com/TanStack/router/commit/a0191afd21afe0e7571af8b0faab171f62e71db7)]:
   - @tanstack/solid-router@2.0.0-alpha.0
 
+## 1.166.15
+
+### Patch Changes
+
+- Updated dependencies [[`91cc628`](https://github.com/TanStack/router/commit/91cc62899b75ca920fe83c5ee7f3dbb5c71a523f)]:
+  - @tanstack/router-core@1.168.1
+  - @tanstack/solid-router@1.168.1
+  - @tanstack/start-client-core@1.167.1
+
+## 1.166.14
+
+### Patch Changes
+
+- Updated dependencies [[`0545239`](https://github.com/TanStack/router/commit/054523900b2ee19308e5a88417dadfc6923afe30)]:
+  - @tanstack/start-client-core@1.167.0
+  - @tanstack/solid-router@1.168.0
+  - @tanstack/router-core@1.168.0
+
 ## 1.166.13
 
 ### Patch Changes